Write undocumented board drawing procedures
Still a bit messy, everything will be documented and a bit rearranged to
be even clearer in the next commit.
For the moment, the setup function initialize the environment for the
GUI, and returns the window Surface object, along with a Clock object.
Then we must call a drawer that take in paramater a draw procedure that
draw everything on the screen. For the moment, this draw procedure also
manage user events, but this will later be splitted to be more concised.
The neat trick inside the drawing of a cell is that, before even
instructing the drawing of a cell, we retrieve from a higher order
function a procedure benifiting from the closure of the earlier, this
closure containing the surface where to draw the cells and the board, we
end up having a draw_cell procedure only needing the coordinates of a
cell to draw the latter.
The board is hardcoded inside the draw function, this is still a
prototype, it will be corrected also in the following commits.