Write docstrings and start redesigning draw flow
draw_board got its parameters inversed, to match the logic of
draw_these_cells.

The flow redesign of the GUI consists into distinguish two orthogonal
parts of the concept of a GUI: the drawings of the GUI and the handling
of the user events. The idea is to have two distinct sub routines that
take care of each these concepts. The computing of the user events will
eventually result in changes that need to be made to the GUI, presented
to the user through its drawing routine. To transmit this potentials
changes to the latter, the sub routine that handles the user events has
to generate a list containing these transformations. This list gets
passed as an argument into the drawing sub routine and there's the magic
happens.

This redesign is not yet implemented but a sketch of it has been done.