Finish board drawing
The way rendering and event computing is done has been thought a bit
more.
Now, everything that shall be drawn on the screen are done through the
use of events. For instance, we now have an event RENDER_BOARD that,
when emitted, will set off a redraw of the board.
Also reordered a bit function definitions to make code more intuitive to
explore.