Exclude diagonals when retrieving adjacent cells
In possible.py.
Before that, get_adjacent_cell was returning all adjacent cells,
including diagonals. But according to the rules of the game we want to
make, a cell is adjacent to the other only when it is to the left, the
right, above or below the latter. That is to say without the diagonals.