Add menu. This made many changes in the flow
Well, the board has been put in a window alonside a menu with three
buttons: score, reset and quit. To manage this, we had to deal with 3
surfaces instead of one. Therefore, they're been put inside a Dict, with
the surface: Window, the subsurface board and the last subsurface menu.
To manage inputs, we call events to launch screen redrawings.